feat: support configuring oci validationType on a rule (#161)
## Description
This PR swaps out configuring the `SkipLayerValidation` flag (which no
longer exists) for the `ValidationType` which can be set to none, fast,
or full.

For now, i've opted to only allow the tui to configure this field on the
host level (ie for all artifacts on in a rule). We can also add extra
prompts to override for a particular artifact, but i felt like it would
just make the tui flow more cumbersome.