feat: allow specifying Azure cloud to connect to (#170)
## Issue
Resolves https://github.com/validator-labs/validatorctl/issues/165.

## Description
Adds support for specifying which Azure cloud to connect to. Users can
use this to enable connecting to Azure Government.

Users are prompted for this when they enable the Azure plugin. Example:

```
Install Azure plugin? [Y/n]: Yes
Select the Azure cloud environment to connect to.
Azure cloud: 
  > AzureCloud
````

Works with both direct rule invocation and installing with Helm.

* For direct rule invocation, it sets the `AZURE_ENVIRONMENT` env var
which gets picked up by the plugin code invoked
* For installing with Helm, it adds the name of the cloud to connect to
to the template file it uses to generate the Helm `values.yaml` file.
This taps into the new Helm config for this in the plugin.

Because this prompt happens before it gets to the rule config, it is
persisted in the Validator config that is generated and users don't have
to select it again when they proceed to the rules step or if they ever
re-configure rules. It's best to tightly couple it to the Validator
config step instead of the rules config step because users would want to
use a particular combination of Tenant ID, Client ID, and Client Secret
when they're connecting to either the public Azure cloud or Azure
Government. It wouldn't make sense to switch from public cloud to Azure
Government but keep the same three credentials. They wouldn't work.

---------

Signed-off-by: Matt Welke <matt.welke@spectrocloud.com>