refactor: update `executePlugins` function to operate on a slice of `PluginSpec`'s (#206)
## Issue
Resolves https://github.com/validator-labs/validatorctl/issues/211

## Description
The purpose of this PR is to refactor the `executePlugin` function to
operate on a slice of `PluginSpec`'s as opposed to a `ValidatorConfig`.
This is prep work to help us soon support the `--crd` flag where users
will be able to pass in CRs for various validation rules.

In the process of making these changes, I updated the oci and network
plugins to fully support direct invocation of rules that previously
required k8s secrets. Now those credentials/pubkeys that were defined in
secrets can be defined inline within the oci and network plugin specs.
To get this fully working, this PR also includes changes to:
- Update OCI prompts to support configuring inline signature
verification public keys
- Update Network prompts to support configuring inline auth details for
HTTPFileRules