refactor: split ConfigureOrCheckCommand into multiple functions (#222)
## Issue
N/A

## Description
Breaks the `ConfigureOrCheckCommand` function down into a
`ConfigureCommand` function and a `CheckCommand` function. Also
introduces a new `configureValidatorConfig` helper function.

With this change, the cognitive load of modifying either the
`ConfigureCommand` or `CheckCommand` functions should be reduced because
we no longer need to keep both direct evaluation and continuous
evaluation in mind.