fix: param name of util function (for linter) (#270)
## Description
Our Revive linter is disallowing the name `len` for a parameter. We made
this code a while ago, but the linter is raising the issue now. It may
be an update to the linter. This changes the parameter name to resolve
it.

Also applies changes from latest `make reviewable` so that the PR can be
merged.

Signed-off-by: Matt Welke <matt.welke@spectrocloud.com>